requirements.txt
vk_api==11.9.1
gTTS==2.2.1
requests==2.25.1
beautifulsoup4==4.9.3
PyYAML==5.4.1